rename aws lambda file
